(move) pure styled componentes to separate files
