+++ src/components/pureStyledComponents/MainScroll.tsx
@@ -1,0 +1,10 @@
+import styled from 'styled-components'
+
+export const MainScroll = styled.div`
+  align-items: center;
+  display: flex;
+  flex-direction: column;
+  flex-grow: 1;
+  overflow: auto;
+  padding-top: 36px;
+`
