+++ src/components/pureStyledComponents/MainWrapper.tsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components'
+
+export const MainWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  height: 100vh;
+  width: 100vw;
+`
