+++ src/index.tsx
@@ -5,7 +5,9 @@ import { ThemeProvider } from 'styled-components'
 
 import { Footer, Header } from './components/layout'
 import { Mainmenu } from './components/navigation/Mainmenu'
-import { InnerContainer, MainScroll, MainWrapper } from './components/pureStyledComponents/layout'
+import { InnerContainer } from './components/pureStyledComponents/InnerContainer'
+import { MainScroll } from './components/pureStyledComponents/MainScroll'
+import { MainWrapper } from './components/pureStyledComponents/MainWrapper'
 import { ApolloProviderWrapper } from './contexts/Apollo'
 import { Web3ContextProvider } from './contexts/Web3Context'
 import { Routes } from './pages/Routes'
