+++ src/pages/condition-detail/ConditionDetailItem.tsx
@@ -2,6 +2,7 @@ import { Provider } from 'ethers/providers'
 import React, { useEffect, useState } from 'react'
 import { CopyToClipboard } from 'react-copy-to-clipboard'
 
+import { Card } from '../../components/common/Card'
 import { INFORMATION_NOT_AVAILABLE } from '../../config/constants'
 import { getKnowOracleFromAddress } from '../../config/networkConfig'
 import { useWeb3Context } from '../../contexts/Web3Context'
@@ -68,7 +69,7 @@ export const ConditionDetailItem = (props: ConditionDetailItemProps) => {
     <>
       {loading && <div>Loading...</div>}
       {!loading && (
-        <>
+        <Card>
           <div className="row">
             <label>Condition Type</label>{' '}
             <label>{isFromOmen ? ConditionType.Omen : ConditionType.Unknown}</label>
@@ -113,7 +114,7 @@ export const ConditionDetailItem = (props: ConditionDetailItemProps) => {
                 truncateStringInTheMiddle(oracle, 6, 6)}
             </label>
           </div>
-        </>
+        </Card>
       )}
     </>
   )
